feat: Modified and fixed rest api methods to make them functional
Modified primitive type properties with wrapper types in Shoe class, renamed "getShoesByCategoria" in "findByCategoria" in ShoeRepository and modified the methods of ShoeService and added PatchMapping in ShoeController

+++ src/main/java/labair_api/services/ShoeService.java
@@ -1,5 +1,7 @@
 package labair_api.services;
 
+import labair_api.exceptions.ExistingShoeException;
+import labair_api.exceptions.ResourceNotFoundException;
 import labair_api.models.Shoe;
 import labair_api.repositories.ShoeRepository;
 import org.springframework.data.domain.Sort;
@@ -11,24 +13,51 @@ import java.util.List;
 public class ShoeService {
     private final ShoeRepository shoeRepository;
 
-    public ShoeService(ShoeRepository shoeRepository) { this.shoeRepository = shoeRepository; }
+    public ShoeService(ShoeRepository shoeRepository) {
+        this.shoeRepository = shoeRepository;
+    }
 
-    public List<Shoe> getAllShoes() {
+    public List<Shoe> findAllShoes() {
         return shoeRepository.findAll();
     }
 
-    public Shoe saveNewShoe(Shoe shoe) {
+    public Shoe createNewShoe(Shoe shoe) {
+        if (shoeRepository.getShoeByNome(shoe.getNome()) != null) {
+            throw new ExistingShoeException();
+        }
+
         return shoeRepository.save(shoe);
     }
 
-    public void deleteShoeById(Long id){
+    public Shoe updateShoeById(Long shoeId, Shoe shoeDetails){
+        Shoe existingShoe = shoeRepository.findById(shoeId).orElseThrow(() -> new ResourceNotFoundException("Scarpa non trovata con id: " + shoeDetails.getId()));
+
+        if(shoeDetails.getNome() != null) existingShoe.setNome(shoeDetails.getNome());
+        if(shoeDetails.getCategoria() != null) existingShoe.setCategoria(shoeDetails.getCategoria());
+        if(shoeDetails.getPrezzo() != null) existingShoe.setPrezzo(shoeDetails.getPrezzo());
+        if(shoeDetails.getTaglieDisponibili() != null) existingShoe.setTaglieDisponibili(shoeDetails.getTaglieDisponibili());
+        if(shoeDetails.getColoriDisponibili() != null) existingShoe.setColoriDisponibili(shoeDetails.getColoriDisponibili());
+        if(shoeDetails.getDescrizione() != null) existingShoe.setDescrizione(shoeDetails.getDescrizione());
+        if(shoeDetails.getImmagineCover() != null) existingShoe.setImmagineCover(shoeDetails.getImmagineCover());
+        if(shoeDetails.getImmaginiScarpa() != null) existingShoe.setImmaginiScarpa(shoeDetails.getImmaginiScarpa());
+        if(shoeDetails.getNuoviArrivi() != null) existingShoe.setNuoviArrivi(existingShoe.getNuoviArrivi());
+        if(shoeDetails.getBestSeller() != null) existingShoe.setBestSeller(existingShoe.getBestSeller());
+
+        return shoeRepository.save(existingShoe);
+    }
+
+    public boolean deleteShoeById(Long id) {
+        if (!shoeRepository.existsById(id)) {
+            return false;
+        }
         shoeRepository.deleteById(id);
+        return true;
     }
 
     public List<Shoe> getShoesByFilters(String category, String sortBy) {
         Sort sortedShoes = Sort.unsorted();
 
-        if(!sortBy.isEmpty()){
+        if (sortBy != null && !sortBy.isEmpty()) {
             sortedShoes = switch (sortBy) {
                 case "newest" -> Sort.by("nuoviArrivi").descending();
                 case "priceAsc" -> Sort.by("prezzo").ascending();
@@ -37,10 +66,17 @@ public class ShoeService {
             };
         }
 
-        return category.isEmpty() ? shoeRepository.findAll(sortedShoes) : shoeRepository.getShoesByCategoria(category, sortedShoes);
+        if(category == null || category.isBlank() || category.equalsIgnoreCase("all"))
+            return shoeRepository.findAll(sortedShoes);
+        else
+            return shoeRepository.findByCategoria(category, sortedShoes);
     }
 
     public Shoe findShoeByNome(String nome) {
+        if (!shoeRepository.existsByNome(nome)) {
+            throw new ResourceNotFoundException("Scarpa non trovata con nome: " + nome);
+        }
+
         return shoeRepository.getShoeByNome(nome);
     }
 }
